+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # GoMLX changelog
 
+# Next
+
+* Added "Flow Matching" examples/demo.
+
 # v0.16.1 - 🎄 2024/12/19 🎄 MatMul fixes
 * MatMul fixed for some edge shape configuration and greatly accelerated in some cases.